requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.14
# by the following command:
#
#    pip-compile requirements.in
#
asgiref==3.11.0
    # via django
django==6.0.1
    # via
    #   -r requirements.in
    #   djangorestframework
djangorestframework==3.16.1
    # via -r requirements.in
pillow==12.1.0
    # via -r requirements.in
sqlparse==0.5.5
    # via django
